BUG FIX: Program losing database connection after some time.
As observed in the last version before this commit, the whole program would break after some time of uptime because the database connection would fail and appear as "Broken pipe" even though there hasn't been any loss of connection.
This commit now implements a 5 connection pool to the server that gets renewed each hour approx. and has less tendency to fail due to this type of connection loss.